Rename ease_functions module to ease_function
