+++ src/curve/ease_function.rs
